LPS-179177 Extract CommerceCountryDetailsScreenNavigationCategory methods to a CommerceCountryDetailsScreenNavigationEntry component and extends CommerceCountryDetailsScreenNavigationCategory in it
